Added Role based access to tickets, fields, users, roles model

+++ src/ticket/ticket.service.ts
@@ -1,7 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { TicketRepository } from './ticket.repository';
 import { CreateTicketDto, UpdateTicketDto } from './dto/ticket.dto';
-import { BaseRepository } from 'src/common/base.repository';
+import { BaseRepository } from 'src/common/repository/base.repository';
 
 @Injectable()
 export class TicketService {
@@ -21,20 +21,37 @@ export class TicketService {
     });
   }
 
-  async getTicketById(id: number) {
-    return this.baseRepository.findById('ticket', id, { fields: true });
+  async getTicketById(id: number, userId: number) {
+    const ticket = await this.baseRepository.findById('ticket', id, {
+      fields: true,
+    });
+    //check user is admin or is trying to get his own ticket
+    const user = await this.baseRepository.findById('user', userId);
+    if ((user && user.roleId === 1) || ticket.assigneeId === userId) {
+      return ticket;
+    }
+    throw new ForbiddenException(
+      'Access denied for performing this operation!',
+    );
   }
 
   async updateTicket(
-    id: string,
+    id: number,
     updateTicketDto: UpdateTicketDto,
     userId: number,
   ) {
     return this.ticketRepository.updateTicket(id, updateTicketDto, userId);
   }
 
-  async deleteTicket(id: number) {
-    return await this.baseRepository.deleteById('ticket', id);
+  async deleteTicket(id: number, userId: number) {
+    const user = await this.baseRepository.findById('user', userId);
+    const ticket = await this.baseRepository.findById('ticket', id);
+    if (ticket.assigneeId === userId || (user && user.roleId === 1)) {
+      return await this.baseRepository.deleteById('ticket', id);
+    }
+    throw new ForbiddenException(
+      'Access denied for performing this operation!',
+    );
   }
 
   async findTicketsByUserId(userId: number) {
